add df streaming

+++ gridfluid.h
@@ -2,6 +2,7 @@
 #define GRIDFLUID_H_INCLUDED
 
 #include <stdint.h>
+#include <stddef.h>
 
 #pragma GCC visibility push(default)
 
@@ -14,15 +15,15 @@ typedef enum e_gridfluid_state {
 
 typedef struct gridfluid *gridfluid_t;
 
-gridfluid_t gridfluid_create_empty_scene(uint8_t x, uint8_t y);
+gridfluid_t gridfluid_create_empty_scene(size_t x, size_t y);
 
 void gridfluid_free(gridfluid_t gf);
 
-void gridfluid_set_obstacle(gridfluid_t gf, uint8_t x, uint8_t y);
-void gridfluid_set_fluid(gridfluid_t gf, uint8_t x, uint8_t y);
-void gridfluid_set_empty(gridfluid_t gf, uint8_t x, uint8_t y);
+void gridfluid_set_obstacle(gridfluid_t gf, size_t x, size_t y);
+void gridfluid_set_fluid(gridfluid_t gf, size_t x, size_t y);
+void gridfluid_set_empty(gridfluid_t gf, size_t x, size_t y);
 void gridfluid_set_gravity(gridfluid_t gf, float g);
-uint8_t gridfluid_get_type(gridfluid_t gf, uint8_t x, uint8_t y);
+uint8_t gridfluid_get_type(gridfluid_t gf, size_t x, size_t y);
 
 void gridfluid_step(gridfluid_t gf);
